style(iter-250): unblock operant-core clippy --fix, 124→28 warnings
cargo clippy --fix --lib -p operant-core had failed outright every prior
session ("failed to automatically apply fixes suggested by rustc").
Root cause: clippy::cmp_owned suggested rewriting
`abs_dir == PathBuf::from("/")` (tools/checkpoint_tool.rs:128) to
`abs_dir == "/"` — but PathBuf doesn't implement PartialEq<&str>, so
that one invalid suggestion broke compilation. cargo clippy --fix rolls
back its entire fix batch for a crate if the post-fix build fails
(all-or-nothing, not per-file), which is why all 96 other legitimate
fixes were silently discarded every time.

Fixed by rewriting the comparison to `abs_dir == Path::new("/")` (valid,
and still avoids the unnecessary allocation clippy was flagging in the
first place). With that one blocker gone, --fix applied cleanly across
39 files: mostly derivable_impls (manual `impl Default` -> `#[derive(Default)]`
+ `#[default]`) and new_without_default, plus manual_div_ceil and a
redundant closure. All behavior-preserving. Ran twice to confirm zero
further auto-fixable suggestions remain.

Remaining 28 warnings are genuine judgment calls (dead_code, non_snake_case
fields tied to external JSON API casing, one type_complexity) — not
attempted here.

All tests pass: 1,603 default + 948 under --features anthropic, 0
regressions.

+++ crates/operant-core/src/kanban/dispatcher.rs
@@ -129,10 +129,6 @@ impl Dispatcher {
                 params![cutoff],
             )
             .map_err(|e| Error::Agent(format!("GC tasks: {}", e)))?;
-        Ok((
-            removed_tasks as usize,
-            removed_runs as usize,
-            removed_events as usize,
-        ))
+        Ok((removed_tasks, removed_runs, removed_events))
     }
 }
